requirements.txt
streamlit
altair
pandas
numpy
openpyxl
google-genai
requests